+++ apps/backend/src/modules/auth/auth.module.ts
@@ -2,12 +2,14 @@ import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { StorageModule } from '../../infrastructure/storage/storage.module';
+import { NotificationModule } from '../notification/notification.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 
 @Module({
   imports: [
     StorageModule,
+    NotificationModule,
     JwtModule.registerAsync({
       global: true,
       imports: [ConfigModule],